electronics_shop/requirements.txt
psycopg2-binary==2.9.11
SQLAlchemy==2.0.44
python-dotenv==1.2.1